Revised control board schematic
Replaced single LDO setup with cascaded LDOs, so that a higher input voltage can be used, and the intermediate 12V can be used to power a fan. Added the appropriate circuitry to control a fan. Added test points for each voltage level and ground.

Layout is still to be updated